requirements.txt
fastapi==0.68.2
uvicorn==0.15.0
sqlalchemy==1.4.25
databases==0.5.2
databases[postgresql]
psycopg2-binary==2.9.1
asyncpg==0.24.0
passlib==1.7.4
pydantic[email]
python-jose==3.3.0